fix: Repository의 인터페이스, 구현체 이름 refactoring
