+++ src/main/java/codesquad/springcafe/repository/user/MemoryUserRepository.java
@@ -2,31 +2,35 @@ package codesquad.springcafe.repository.user;
 
 import codesquad.springcafe.domain.User;
 import codesquad.springcafe.dto.UserUpdateDto;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
 
-@Primary
 @Repository
-public class H2UserRepository implements  UserRepositoryInterface {
+public class MemoryUserRepository implements UserRepository {
+    private static final List<User> users = new ArrayList<>();
+
     @Override
     public User createUser(User user) {
-        return null;
+        users.add(user);
+        return user;
     }
 
     @Override
     public User updateUser(String userId, UserUpdateDto userUpdateDto) {
-        return null;
+        User user = findByUserId(userId).get();
+        user.update(userUpdateDto);
+        return user;
     }
 
     @Override
     public List<User> findAllUsers() {
-        return null;
+        return users;
     }
 
     @Override
     public Optional<User> findByUserId(String userId) {
-        return Optional.empty();
+        return users.stream().filter(user -> user.getUserId().equals(userId)).findFirst();
     }
 }
